renamed quicklook.vue to Quicklook.vue
